LPS-201494 Rename
