+++ modules/apps/segments/segments-web/src/main/resources/META-INF/resources/js/components/criteria_builder/CriteriaGroup.js
@@ -14,7 +14,7 @@ import {
 	SUPPORTED_OPERATORS,
 	SUPPORTED_PROPERTY_TYPES,
 } from '../../utils/constants';
-import {DragTypes} from '../../utils/drag-types';
+import {DragTypes} from '../../utils/dragTypes';
 import {
 	generateGroupId,
 	getChildGroupIds,
