+++ modules/apps/segments/segments-web/src/main/resources/META-INF/resources/js/components/criteria_builder/DropZone.js
@@ -8,7 +8,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import {DropTarget as dropTarget} from 'react-dnd';
 
-import {DragTypes} from '../../utils/drag-types';
+import {DragTypes} from '../../utils/dragTypes';
 
 const {CRITERIA_GROUP, CRITERIA_ROW, PROPERTY} = DragTypes;
