+++ modules/apps/segments/segments-web/src/main/resources/META-INF/resources/js/utils/dragTypes.ts
